feat(v1beta4): convert test to be compliant to v1beta4
