+++ charts/v0.5.0/values.yaml
@@ -75,7 +75,7 @@ providers:
     image:
       prefix: ghcr.io/syngit-org
       name: syngit-provider-gitlab
-      tag: v0.2.1
+      tag: v0.2.3
       # imagePullSecrets:
       # imagePullPolicy:
 
@@ -84,6 +84,6 @@ providers:
     image:
       prefix: ghcr.io/syngit-org
       name: syngit-provider-github
-      tag: v0.1.0
+      tag: v0.1.2
       # imagePullSecrets:
       # imagePullPolicy:
