+++ config/crd/bases/syngit.io_remoteuserbindings.yaml
@@ -536,3 +536,5 @@ spec:
         type: object
     served: true
     storage: true
+    subresources:
+      status: {}
