+++ config/crd/bases/syngit.io_remoteusers.yaml
@@ -292,3 +292,5 @@ spec:
         type: object
     served: true
     storage: true
+    subresources:
+      status: {}
