+++ pkg/api/v1beta1/remotesyncer_conversion.go
@@ -17,12 +17,12 @@ limitations under the License.
 package v1beta1
 
 import (
-	v1beta3 "github.com/syngit-org/syngit/pkg/api/v1beta3"
+	v1beta4 "github.com/syngit-org/syngit/pkg/api/v1beta4"
 	"sigs.k8s.io/controller-runtime/pkg/conversion"
 )
 
 func (src *RemoteSyncer) ConvertTo(dstRaw conversion.Hub) error {
-	dst := dstRaw.(*v1beta3.RemoteSyncer)
+	dst := dstRaw.(*v1beta4.RemoteSyncer)
 
 	// Common conversion
 	dst.ObjectMeta = src.ObjectMeta
@@ -30,14 +30,14 @@ func (src *RemoteSyncer) ConvertTo(dstRaw conversion.Hub) error {
 	dst.Spec.DefaultBranch = src.Spec.DefaultBranch
 	dst.Spec.BypassInterceptionSubjects = src.Spec.BypassInterceptionSubjects
 	dst.Spec.DefaultBlockAppliedMessage = src.Spec.DefaultBlockAppliedMessage
-	dst.Spec.DefaultUnauthorizedUserMode = v1beta3.DefaultUnauthorizedUserMode(src.Spec.DefaultUnauthorizedUserMode)
+	dst.Spec.DefaultUnauthorizedUserMode = v1beta4.DefaultUnauthorizedUserMode(src.Spec.DefaultUnauthorizedUserMode)
 	dst.Spec.DefaultRemoteUserRef = src.Spec.DefaultRemoteUserRef
 	dst.Spec.ExcludedFields = src.Spec.ExcludedFields
 	dst.Spec.ExcludedFieldsConfigMapRef = src.Spec.ExcludedFieldsConfigMapRef
 	dst.Spec.RemoteRepository = src.Spec.RemoteRepository
 	dst.Spec.RootPath = src.Spec.RootPath
-	dst.Spec.ScopedResources = v1beta3.ScopedResources(src.Spec.ScopedResources)
-	dst.Spec.Strategy = v1beta3.Strategy(src.Spec.ProcessMode)
+	dst.Spec.ScopedResources = v1beta4.ScopedResources(src.Spec.ScopedResources)
+	dst.Spec.Strategy = v1beta4.Strategy(src.Spec.ProcessMode)
 	dst.Spec.CABundleSecretRef = src.Spec.CABundleSecretRef
 
 	// Target transfer
@@ -45,17 +45,17 @@ func (src *RemoteSyncer) ConvertTo(dstRaw conversion.Hub) error {
 		dst.Annotations = map[string]string{}
 	}
 	if string(src.Spec.PushMode) == string(SameBranch) {
-		dst.Spec.TargetStrategy = v1beta3.OneTarget
-		dst.Annotations[v1beta3.RtAnnotationKeyOneOrManyBranches] = src.Spec.DefaultBranch
+		dst.Spec.TargetStrategy = v1beta4.OneTarget
+		dst.Annotations[v1beta4.RtAnnotationKeyOneOrManyBranches] = src.Spec.DefaultBranch
 	} else {
-		dst.Spec.TargetStrategy = v1beta3.MultipleTarget
+		dst.Spec.TargetStrategy = v1beta4.MultipleTarget
 	}
 
 	return nil
 }
 
 func (dst *RemoteSyncer) ConvertFrom(srcRaw conversion.Hub) error {
-	src := srcRaw.(*v1beta3.RemoteSyncer)
+	src := srcRaw.(*v1beta4.RemoteSyncer)
 
 	// Common conversion
 	dst.ObjectMeta = src.ObjectMeta
@@ -74,7 +74,7 @@ func (dst *RemoteSyncer) ConvertFrom(srcRaw conversion.Hub) error {
 	dst.Spec.CABundleSecretRef = src.Spec.CABundleSecretRef
 
 	// Target transfer
-	if src.Spec.TargetStrategy == v1beta3.OneTarget {
+	if src.Spec.TargetStrategy == v1beta4.OneTarget {
 		dst.Spec.PushMode = SameBranch
 	} else {
 		dst.Spec.PushMode = MultipleBranch
