+++ pkg/api/v1beta3/remoteuser_conversion.go
@@ -16,4 +16,55 @@ limitations under the License.
 
 package v1beta3
 
-func (*RemoteUser) Hub() {}
+import (
+	"github.com/syngit-org/syngit/pkg/api/v1beta4"
+	"sigs.k8s.io/controller-runtime/pkg/conversion"
+)
+
+func (src *RemoteUser) ConvertTo(dstRaw conversion.Hub) error {
+	dst := dstRaw.(*v1beta4.RemoteUser)
+
+	// Common conversion
+	dst.ObjectMeta = src.ObjectMeta
+
+	dst.Spec.Email = src.Spec.Email
+	gitBaseDomainFQDN := src.Spec.GitBaseDomainFQDN
+	dst.Spec.GitBaseDomainFQDN = gitBaseDomainFQDN
+	dst.Spec.SecretRef = src.Spec.SecretRef
+
+	// Status conversion - preserve existing status from hub
+	// Only convert if src has status set (for backward compatibility)
+	if len(src.Status.Conditions) > 0 || src.Status.SecretBoundStatus != "" {
+		dst.Status.Conditions = src.Status.Conditions
+		dst.Status.ConnexionStatus.Details = src.Status.ConnexionStatus.Details
+		dst.Status.ConnexionStatus.Status = v1beta4.RemoteUserConnexionStatusReason(src.Status.ConnexionStatus.Status)
+		dst.Status.GitUser = src.Status.GitUser
+		dst.Status.SecretBoundStatus = v1beta4.SecretBoundStatus(src.Status.SecretBoundStatus)
+	}
+
+	return nil
+}
+
+func (dst *RemoteUser) ConvertFrom(srcRaw conversion.Hub) error {
+	src := srcRaw.(*v1beta4.RemoteUser)
+
+	// Common conversion
+	dst.ObjectMeta = src.ObjectMeta
+
+	dst.Spec.Email = src.Spec.Email
+	gitBaseDomainFQDN := src.Spec.GitBaseDomainFQDN
+	dst.Spec.GitBaseDomainFQDN = gitBaseDomainFQDN
+	dst.Spec.SecretRef = src.Spec.SecretRef
+
+	// Status conversion - preserve existing status from hub
+	// Only convert if src has status set (for backward compatibility)
+	if len(src.Status.Conditions) > 0 || src.Status.SecretBoundStatus != "" {
+		dst.Status.Conditions = src.Status.Conditions
+		dst.Status.ConnexionStatus.Details = src.Status.ConnexionStatus.Details
+		dst.Status.ConnexionStatus.Status = RemoteUserConnexionStatusReason(src.Status.ConnexionStatus.Status)
+		dst.Status.GitUser = src.Status.GitUser
+		dst.Status.SecretBoundStatus = SecretBoundStatus(src.Status.SecretBoundStatus)
+	}
+
+	return nil
+}
