+++ pkg/api/v1beta3/remoteuserbinding_conversion.go
@@ -16,4 +16,33 @@ limitations under the License.
 
 package v1beta3
 
-func (*RemoteUserBinding) Hub() {}
+import (
+	"github.com/syngit-org/syngit/pkg/api/v1beta4"
+	"sigs.k8s.io/controller-runtime/pkg/conversion"
+)
+
+func (src *RemoteUserBinding) ConvertTo(dstRaw conversion.Hub) error {
+	dst := dstRaw.(*v1beta4.RemoteUserBinding)
+
+	// Common conversion
+	dst.ObjectMeta = src.ObjectMeta
+
+	dst.Spec.RemoteUserRefs = src.Spec.RemoteUserRefs
+	dst.Spec.RemoteTargetRefs = src.Spec.RemoteTargetRefs
+	dst.Spec.Subject = src.Spec.Subject
+
+	return nil
+}
+
+func (dst *RemoteUserBinding) ConvertFrom(srcRaw conversion.Hub) error {
+	src := srcRaw.(*v1beta4.RemoteUserBinding)
+
+	// Common conversion
+	dst.ObjectMeta = src.ObjectMeta
+
+	dst.Spec.RemoteUserRefs = src.Spec.RemoteUserRefs
+	dst.Spec.RemoteTargetRefs = src.Spec.RemoteTargetRefs
+	dst.Spec.Subject = src.Spec.Subject
+
+	return nil
+}
