+++ pkg/api/v1beta4/remoteuserbinding_types.go
@@ -72,15 +72,15 @@ type RemoteUserBindingStatus struct {
 	UserKubernetesID string `json:"userKubernetesID,omitempty" protobuf:"bytes,4,rep,name=userKubernetesID"`
 }
 
-// +kubebuilder:subresource:status
 // +kubebuilder:resource:path=remoteuserbindings,shortName=rub;rubs,categories=syngit
 
 // +kubebuilder:printcolumn:name="Kubernetes User ID",type=string,JSONPath=`.status.userKubernetesID`,priority=0
 // +kubebuilder:printcolumn:name="Remote Users State",type=string,JSONPath=`.status.remoteUserState`,priority=1
 // +kubebuilder:printcolumn:name="Age",type=string,JSONPath=`.metadata.creationTimestamp`,priority=0
 
-// +kubebuilder:storageversion
 // +kubebuilder:object:root=true
+// +kubebuilder:subresource:status
+// +kubebuilder:storageversion
 
 // RemoteUserBinding is the Schema for the remoteuserbindings API
 type RemoteUserBinding struct {
