+++ test/e2e/samples/syngit_v1beta3_remotesyncer.yaml
