+++ test/e2e/samples/syngit_v1beta4_remotesyncer.yaml
